# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Project API",
        "description": "Badminton Project API Documentation",
        "contact": {
            "email": "admin@badminton.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Server"
        },
        {
            "url": "https://api.badminton.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/getBookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Lấy danh sách sân đã đặt của người dùng",
                "description": "API lấy danh sách sân đã đặt của một user với status linh hoạt",
                "operationId": "getBookings",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái của booking (awaiting, accepted, declined, cancelled)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "awaiting",
                                    "accepted",
                                    "declined",
                                    "cancelled"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/bookCourt": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Đặt sân",
                "description": "API để đặt sân badminton",
                "operationId": "bookCourt",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "venue_id",
                                    "user_id",
                                    "booking_date",
                                    "courts"
                                ],
                                "properties": {
                                    "venue_id": {
                                        "description": "ID của sân",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID của người đặt",
                                        "type": "integer"
                                    },
                                    "booking_date": {
                                        "description": "Ngày đặt (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "courts": {
                                        "description": "Danh sách sân cần đặt",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "court_number": {
                                                    "description": "Số sân",
                                                    "type": "integer"
                                                },
                                                "start_time": {
                                                    "description": "Thời gian bắt đầu (HH:MM)",
                                                    "type": "string",
                                                    "format": "time"
                                                },
                                                "end_time": {
                                                    "description": "Thời gian kết thúc (HH:MM)",
                                                    "type": "string",
                                                    "format": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payment_image": {
                                        "description": "Ảnh chuyển khoản dạng base64 (tuỳ chọn)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đặt sân thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking successful"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lấy danh sách tất cả người dùng",
                "description": "API lấy tất cả người dùng kèm thông tin rating",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xoá người dùng",
                "description": "API xoá người dùng",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng cần xoá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xoá thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xóa người dùng thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}